Remove obsoleted stuff about SynthesizerState
