+++ src/framework/audio/internal/abstractsynthesizer.h
@@ -28,7 +28,6 @@
 #include "mpe/events.h"
 #include "modularity/ioc.h"
 
-#include "../synthtypes.h"
 #include "../audiotypes.h"
 #include "../isynthesizer.h"
 #include "../iaudioconfiguration.h"
